backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# LangChain Core
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-chroma==0.1.4
langchain-community==0.3.13
langchain-docling==0.2.0

# Document Processing
docling==2.14.0
docling-core==2.7.1

# Vector Store & Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Cross-Encoder for Reranking
cross-encoder==0.1.0

# arXiv API
arxiv==2.1.3

# Environment & Configuration
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.4
pydantic-settings==2.7.0
